// tests/sanity/tests/enums.ts
export enum Columns {
    Employee = 'Employee',
    Person = 'Person',
    Company = 'Company',
    Location = 'Location',
    Attachments = 'Attachments',
    ModifiedDate = 'Modified date',
    Role = 'Role',
    ContactInfo = 'Contact Info',
    ModifiedBy = 'Modified by',
    CreatedBy = 'Created by',
    CreatedDate = 'Created date',
    Birthday = 'Birthday',
    Department = 'Department',
    Title = 'Title',
    Onsite = 'Onsite',
    Remote = 'Remote',
    Source = 'Source',
    Description = 'Description'
  }

  export enum ContactType {
    Employee = 'Employee',
    Person = 'Person',
    Company = 'Company'
  }